# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Era Task API Documentation",
        "description": "API documentation for authentication and admin management.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/admin/register": {
            "post": {
                "tags": [
                    "Admin Authentication"
                ],
                "summary": "Register a new admin",
                "operationId": "82eef8879ce848a3d6b25cee980b9cdd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterAdminRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegisterAdminResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/login": {
            "post": {
                "tags": [
                    "Admin Authentication"
                ],
                "summary": "Login an existing admin user",
                "operationId": "fb63ce6528ee8d3c5a1188b683e05552",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginAdminRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginAdminResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/logout": {
            "post": {
                "tags": [
                    "Admin Authentication"
                ],
                "summary": "Logout the authenticated admin",
                "operationId": "aae3f98888542465efe98904b0f20087",
                "responses": {
                    "200": {
                        "description": "Successful logout",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LogoutAdminResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LogoutErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "LoginAdminRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "admin@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password123"
                    }
                },
                "type": "object"
            },
            "LoginAdminResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "token": {
                        "type": "string",
                        "example": "1|abcde12345"
                    },
                    "data": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Sadig Bahlulzada"
                            },
                            "email": {
                                "type": "string",
                                "example": "admin@example.com"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "LoginErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Invalid credentials"
                    }
                },
                "type": "object"
            },
            "LogoutAdminResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "LogoutErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "RegisterAdminRequest": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Sadig Bahlulzada"
                    },
                    "email": {
                        "type": "string",
                        "example": "admin@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password123"
                    }
                },
                "type": "object"
            },
            "RegisterAdminResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "token": {
                        "type": "string",
                        "example": "1|abcde12345"
                    },
                    "data": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Sadig Bahlulzada"
                            },
                            "email": {
                                "type": "string",
                                "example": "admin@example.com"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Something went wrong"
                    },
                    "error": {
                        "type": "string",
                        "example": "Server Error"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "tokenUrl": "http://127.0.0.1:8000/oauth/token",
                        "refreshUrl": "http://127.0.0.1:8000oauth/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}